requirements.txt
openai==1.2.3
rich==13.6.0
requests==2.31.0
pygithub==2.1.1
serpapi==0.1.5
python-dotenv==1.0.0
